chore(v3): clean up resource plugin folder (#6335)
* refactor: clean up resource plugin folder

* fix: ut
