+++ packages/fx-core/tests/core/other.test.ts
@@ -39,7 +39,7 @@ import {
 import {
   upgradeDefaultFunctionName,
   upgradeProgrammingLanguage,
-} from "../../src/core/middleware/envInfoLoader";
+} from "../../src/core/middleware/envInfoLoaderV3";
 import { createAppNameQuestion } from "../../src/core/question";
 import { resourceGroupHelper } from "../../src/plugins/solution/fx-solution/utils/ResourceGroupHelper";
 import { parseTeamsAppTenantId } from "../../src/plugins/solution/fx-solution/v2/utils";
